+++ src/main/java/Duke.java
@@ -1,7 +1,6 @@
 import java.io.IOException;
-import java.util.Scanner; //Import Java Scanner
-import java.util.ArrayList; //Import Java ArrayList
-import java.io.BufferedReader;
+import java.util.Scanner;
+import java.util.ArrayList;
 
 public class Duke {
     //Create task ArrayList
@@ -10,39 +9,17 @@ public class Duke {
     //Declare constant variables
     private static final String CHAR_FALSE = "0";
     private static final String CHAR_TRUE = "1";
+
     private static final String CHAR_TODO = "T";
     private static final String CHAR_DEADLINE = "D";
     private static final String CHAR_EVENT = "E";
-    private static final String STRING_BYE = "bye";
-    private static final String STRING_LIST = "list";
-    private static final String STRING_DONE = "done";
-    private static final String STRING_DELETE = "delete";
-    private static final String STRING_TODO = "todo";
-    private static final String STRING_DEADLINE = "deadline";
-    private static final String STRING_EVENT = "event";
-    private static final String LINE_SEPARATOR = "____________________________________________________________\n";
-    private static final String MSG_GREET = LINE_SEPARATOR + "Hello! I'm Duke\n" + "What can I do for you?\n" + LINE_SEPARATOR;
-    private static final String MSG_BYE = LINE_SEPARATOR + "Bye. Hope to see you again soon!\n" + LINE_SEPARATOR;
-    private static final String MSG_LIST = LINE_SEPARATOR + "Here are the tasks in your list:";
-    private static final String MSG_DONE = LINE_SEPARATOR + "Nice! I've marked this task as done:";
-    private static final String MSG_DELETE = LINE_SEPARATOR + "Noted. I've removed this task:";
-    private static final String MSG_PRE_TASK = LINE_SEPARATOR + "Got it. I've added this task:";
-    private static final String MSG_POST_TASK_1 = "Now you have ";
-    private static final String MSG_POST_TASK_2 = " tasks in the list.\n" + LINE_SEPARATOR;
-    private static final String ERROR_MSG_IO = "Input or output failure!\n";
 
-    //Store keywords' number of characters
-    private static int doneStrLen = STRING_DONE.length();
-    private static int deleteStrLen = STRING_DELETE.length();
-    private static int todoStrLen = STRING_TODO.length();
-    private static int deadlineStrLen = STRING_DEADLINE.length();
-    private static int eventStrLen = STRING_EVENT.length();
+    //Create class
+    private static Ui ui;
 
-    //Declare error inputs in an array
-    private static String[] errorInputList = {"blah"};
-
-    public static void greetUser() {
-        System.out.print(MSG_GREET);
+    //Constructor
+    public Duke() {
+        ui = new Ui();
     }
 
     public static void appendTaskToFile(Storage appStorage, String filePath, Task currentTask, String errorMessage) {
@@ -190,10 +167,7 @@ public class Duke {
                 tasks.add(new Event(inputExtract) );
                 appendTaskToFile(appStorage, filePath, tasks.get(tasks.size() -1), ERROR_MSG_IO);
             } else {
-                //Add task
-                inputExtract = input;
-                tasks.add(new Task(inputExtract) );
-                appendTaskToFile(appStorage, filePath, tasks.get(tasks.size() -1), ERROR_MSG_IO);
+                throw new DukeException();
             }
 
             System.out.println(MSG_PRE_TASK);
