+++ src/main/java/Todo.java
@@ -2,6 +2,6 @@ public class Todo extends Task {
     //Constructor
     public Todo(String taskDescription) {
         super(taskDescription);
-        typeIdt = 'T';
+        typeIdt = CHAR_TODO;
     }
 }
